pipevision/requirements.txt
# PipeVision Requirements
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
geoalchemy2>=0.14.0
alembic>=1.13.0

# Redis & Celery
celery>=5.3.0
redis>=5.0.0

# DWG/DXF Processing
ezdxf>=1.1.0

# Geospatial
pyproj>=3.6.0
shapely>=2.0.0

# Image Processing
Pillow>=10.2.0

# 3D Export
trimesh>=4.0.0
numpy>=1.26.0

# HTTP Client (for webhooks)
httpx>=0.26.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0

# Utilities
python-dateutil>=2.8.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0

# Production
gunicorn>=21.2.0
